Pass SSL options to git through the environment
Because the git executable is also used to access the remote github
endpoint we need to pass it the SSL configuration options. Since git
supports only a subset of the request library SSL options, that is the
only options we support.

The PATH is mirrored to the exec'd git so the user specified git
executable is used. This is important because by default on OS X curl
tries to use the OS X keychain for SSL certificate loading/lookup, but
that leads to a conflict in configuration format. OS X curl will try to
import an explicitly given certificate into the keychain, but it lacks
the ability to detect this has already been done and will error out with
a duplicate certificate error. Instead you should specify the
certificate name in the keychain to be used, but request lacks the
ability to use that information.

On OS X the simplest way to have git work with client SSL certificates
is to install git using homebrew using the --with-brewed-curl option and
to install curl first using --with-openssl or any other SSL
implementation other than Secure Transport.